requirements.txt
freeopcua 
sqlalchemy
flask
typing-extensions